Move StateListenerBeforeApply
